Product_Sales_Predictions_and_Segementation
Got the real time dataset and done some initial analysis in MS-Excel. Loaded the dataset in Python.
Understood the dataset. Checked for duplicate rows and columns. Checked and treated the Outliers and Null
values. Plotted graphs and drawn insights. Done Feature Engineering and Segmentation. Encoded the
categorical variables and scaled the data. Separated the train and test set for model building. Built the base
model as linear regression and advance model as Gradient Boosting Regressor.